3 of a kind 5 hand works

+++ ADSSE miniproject poker prob/ADSSE miniproject poker prob/BuildDeck.cs
@@ -46,9 +46,9 @@ namespace ADSSE_miniproject_poker_prob
             return cards;
         }
 
-        public int CardsLeft(List<Card> deck)
+        public float CardsLeft(List<Card> deck)
         {
-            int amountOfCardsLeft = 0;
+            float amountOfCardsLeft = 0;
 
             for(int c = 0; c < deck.Count; c++)
             {
@@ -57,9 +57,9 @@ namespace ADSSE_miniproject_poker_prob
             return amountOfCardsLeft;
         }
 
-        public int TypeOfCardLeft(List<Card> deck, int i)
+        public float TypeOfCardLeft(List<Card> deck, int i)
         {
-            int typeLeft = 0;
+            float typeLeft = 0;
             for (int a = 0; a < deck.Count; a++)
             {
                 if (deck[a].rank == i && deck[a].available) typeLeft += 1;
@@ -67,9 +67,9 @@ namespace ADSSE_miniproject_poker_prob
             return typeLeft;
         }
 
-        public int AcesLeft(List<Card> deck)
+        public float AcesLeft(List<Card> deck)
         {
-            int acesLeft = 0;
+            float acesLeft = 0;
             for (int a = 0; a < deck.Count; a++)
             {
                 if(deck[a].rank == 0 && deck[a].available) acesLeft += 1;
@@ -77,9 +77,9 @@ namespace ADSSE_miniproject_poker_prob
             return acesLeft;
         }
 
-        public int TwosLeft(List<Card> deck)
+        public float TwosLeft(List<Card> deck)
         {
-            int twosLeft = 0;
+            float twosLeft = 0;
             for(int tw = 0; tw < deck.Count; tw++)
             {
                 if (deck[tw].rank == 1 && deck[tw].available) twosLeft += 1;
@@ -87,9 +87,9 @@ namespace ADSSE_miniproject_poker_prob
             return twosLeft;
         }
 
-        public int ThreesLeft(List<Card> deck)
+        public float ThreesLeft(List<Card> deck)
         {
-            int threesLeft = 0;
+            float threesLeft = 0;
             for(int th = 0; th < deck.Count; th++)
             {
                 if (deck[th].rank == 2 && deck[th].available) threesLeft += 1;
@@ -97,9 +97,9 @@ namespace ADSSE_miniproject_poker_prob
             return threesLeft;
         }
 
-        public int FoursLeft(List<Card> deck)
+        public float FoursLeft(List<Card> deck)
         {
-            int foursLeft = 0;
+            float foursLeft = 0;
             for (int fo = 0; fo < deck.Count; fo++)
             {
                 if (deck[fo].rank == 3 && deck[fo].available) foursLeft += 1; 
@@ -107,9 +107,9 @@ namespace ADSSE_miniproject_poker_prob
             return foursLeft;
         }
 
-        public int FivesLeft(List<Card> deck)
+        public float FivesLeft(List<Card> deck)
         {
-            int fivesLeft = 0;
+            float fivesLeft = 0;
             for(int fv = 0; fv < deck.Count; fv++)
             {
                 if (deck[fv].rank == 4 && deck[fv].available) fivesLeft += 1;
@@ -117,9 +117,9 @@ namespace ADSSE_miniproject_poker_prob
             return fivesLeft;
         }
 
-        public int SixsLeft(List<Card> deck)
+        public float SixsLeft(List<Card> deck)
         {
-            int sixsLeft = 0;
+            float sixsLeft = 0;
             for(int si = 0; si < deck.Count; si++)
             {
                 if (deck[si].rank == 5 && deck[si].available) sixsLeft += 1;
@@ -127,9 +127,9 @@ namespace ADSSE_miniproject_poker_prob
             return sixsLeft;
         }
 
-        public int SevensLeft(List<Card> deck)
+        public float SevensLeft(List<Card> deck)
         {
-            int sevensLeft = 0;
+            float sevensLeft = 0;
             for(int sv = 0; sv < deck.Count; sv++)
             {
                 if (deck[sv].rank == 6 && deck[sv].available) sevensLeft += 1;
@@ -137,9 +137,9 @@ namespace ADSSE_miniproject_poker_prob
             return sevensLeft;
         }
 
-        public int EightsLeft(List<Card> deck)
+        public float EightsLeft(List<Card> deck)
         {
-            int eightsLeft = 0;
+            float eightsLeft = 0;
             for (int e = 0; e < deck.Count; e++)
             {
                 if (deck[e].rank == 7 && deck[e].available) eightsLeft += 1;
@@ -147,9 +147,9 @@ namespace ADSSE_miniproject_poker_prob
             return eightsLeft;
         }
 
-        public int NinesLeft(List<Card> deck)
+        public float NinesLeft(List<Card> deck)
         {
-            int ninesLeft = 0;
+            float ninesLeft = 0;
             for (int n = 0; n < deck.Count; n++)
             {
                 if (deck[n].rank == 8 && deck[n].available) ninesLeft += 1;
@@ -157,9 +157,9 @@ namespace ADSSE_miniproject_poker_prob
             return ninesLeft;
         }
 
-        public int TensLeft(List<Card> deck)
+        public float TensLeft(List<Card> deck)
         {
-            int tensLeft = 0;
+            float tensLeft = 0;
             for (int ten = 0; ten < deck.Count; ten++)
             {
                 if (deck[ten].rank == 9 && deck[ten].available) tensLeft += 1;
@@ -167,9 +167,9 @@ namespace ADSSE_miniproject_poker_prob
             return tensLeft;
         }
 
-        public int JacksLeft(List<Card> deck)
+        public float JacksLeft(List<Card> deck)
         {
-            int jacksLeft = 0;
+            float jacksLeft = 0;
             for (int j = 0; j < deck.Count; j++)
             {
                 if (deck[j].rank == 10 && deck[j].available) jacksLeft += 1;
@@ -177,9 +177,9 @@ namespace ADSSE_miniproject_poker_prob
             return jacksLeft;
         }
 
-        public int QueensLeft(List<Card> deck)
+        public float QueensLeft(List<Card> deck)
         {
-            int queensLeft = 0;
+            float queensLeft = 0;
             for (int q = 0; q < deck.Count; q++)
             {
                 if (deck[q].rank == 11 && deck[q].available) queensLeft += 1;
@@ -187,9 +187,9 @@ namespace ADSSE_miniproject_poker_prob
             return queensLeft;
         }
 
-        public int KingsLeft(List<Card> deck)
+        public float KingsLeft(List<Card> deck)
         {
-            int kingsLeft = 0;
+            float kingsLeft = 0;
             for (int k = 0; k < deck.Count; k++)
             {
                 if (deck[k].rank == 12 && deck[k].available) kingsLeft += 1;
@@ -197,9 +197,9 @@ namespace ADSSE_miniproject_poker_prob
             return kingsLeft;
         }
 
-        public int HeartsLeft(List<Card> deck)
+        public float HeartsLeft(List<Card> deck)
         {
-            int heartlsLeft = 0;
+            float heartlsLeft = 0;
             for (int hl = 0; hl < deck.Count; hl++)
             {
                 if (deck[hl].suit == 0 && deck[hl].available) heartlsLeft += 1;
@@ -207,9 +207,9 @@ namespace ADSSE_miniproject_poker_prob
             return heartlsLeft;
         }
 
-        public int ClubsLeft(List<Card> deck)
+        public float ClubsLeft(List<Card> deck)
         {
-            int clubsLeft = 0;
+            float clubsLeft = 0;
             for (int cl = 0; cl < deck.Count; cl++)
             {
                 if (deck[cl].suit == 1 && deck[cl].available) clubsLeft += 1;
@@ -217,9 +217,9 @@ namespace ADSSE_miniproject_poker_prob
             return clubsLeft;
         }
 
-        public int DiamondsLeft(List<Card> deck)
+        public float DiamondsLeft(List<Card> deck)
         {
-            int diamondsLeft = 0;
+            float diamondsLeft = 0;
             for (int dl = 0; dl < deck.Count; dl++)
             {
                 if (deck[dl].suit == 2 && deck[dl].available) diamondsLeft += 1;
@@ -227,9 +227,9 @@ namespace ADSSE_miniproject_poker_prob
             return diamondsLeft;
         }
 
-        public int SpadesLeft(List<Card> deck)
+        public float SpadesLeft(List<Card> deck)
         {
-            int spadesLeft = 0;
+            float spadesLeft = 0;
             for (int sl = 0; sl < deck.Count; sl++)
             {
                 if (deck[sl].suit == 3 && deck[sl].available) spadesLeft += 1;
